+++ app/router.scrollBehavior.js
@@ -1,12 +1,31 @@
-export default function (to, from, savedPosition) {
-  if (to.hash) {
-    return window.scrollTo({
-      top:
-        (document.querySelector(to.hash)
-          ? document.querySelector(to.hash).offsetTop
-          : 0) + window.innerHeight,
-      behavior: 'smooth',
-    })
+export default async function (to, from, savedPosition) {
+  if (savedPosition) {
+    return savedPosition
   }
-  return window.scrollTo({ top: 0, behavior: 'smooth' })
+
+  const findEl = async (hash, x = 0) => {
+    const id = String(hash).replace('#', '')
+    return (
+      document.getElementById(id) ||
+      (await new Promise((resolve) => {
+        if (x > 50) {
+          return resolve(document.querySelector('#app'))
+        }
+        setTimeout(() => {
+          resolve(findEl(hash, ++x || 1))
+        }, 100)
+      }))
+    )
+  }
+
+  if (to.hash) {
+    const el = await findEl(to.hash)
+    if ('scrollBehavior' in document.documentElement.style) {
+      return window.scrollTo({ top: el.offsetTop, behavior: 'smooth' })
+    } else {
+      return window.scrollTo(0, el.offsetTop)
+    }
+  }
+
+  return { x: 0, y: 0 }
 }