+++ store/referrals.js
@@ -3,6 +3,7 @@
 
 export const state = () => ({
   list: [],
+  filteredList: [],
 })
 
 export const mutations = {
@@ -12,6 +13,14 @@ export const mutations = {
 }
 
 export const actions = {
+  async findAll({ commit, dispatch, getters }, slug) {
+    await dispatch('all')
+    let data = getters.list
+    data = data
+      .filter((b) => b.community)
+      .filter((b) => b.community.slug === slug)
+    commit('setFilteredList', data)
+  },
   async all({ commit }) {
     const { data } = await this.$api.get('referrals/list')
     commit('setList', data)
@@ -21,5 +30,8 @@ export const actions = {
 export const getters = {
   list(state) {
     return state.list
+  },
+  filteredList(state) {
+    return state.filteredList
   },
 }