+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.6.5](https://gitlab.com/u2467/dacade/dacade/compare/v2.6.4...v2.6.5) (2021-11-09)
+
+
+### Bug Fixes
+
+* bounty card link ([d3b14d0](https://gitlab.com/u2467/dacade/dacade/commit/d3b14d0298b76d38f8694d7e201cc4321498d6e4))
+
+### [2.6.4](https://gitlab.com/u2467/dacade/dacade/compare/v2.6.3...v2.6.4) (2021-11-08)
+
+
+### Features
+
+* add markdown heading navigation ([75d2d23](https://gitlab.com/u2467/dacade/dacade/commit/75d2d2395bec4612e51b125458fe1fdd22e69031))
+
+### [2.6.3](https://gitlab.com/u2467/dacade/dacade/compare/v2.6.2...v2.6.3) (2021-11-08)
+
+
+### Features
+
+* show bounties for each reward ([0d1e70d](https://gitlab.com/u2467/dacade/dacade/commit/0d1e70d3b045d297084609cb9af25259620046a8))
+
+
+### Bug Fixes
+
+* filtered list ([dd05693](https://gitlab.com/u2467/dacade/dacade/commit/dd056935a80b5160272b733a60ec9d94b1ea9fe6))
+
 ### [2.6.2](https://gitlab.com/u2467/dacade/dacade/compare/v2.6.1...v2.6.2) (2021-11-03)